+++ config/gatsby-browser.tsx
@@ -17,10 +17,17 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
+import { Global } from '@emotion/react'
 import React from 'react'
 import { I18nextProvider } from 'react-i18next'
 import i18next from '../src/lib/i18next'
 import '../src/utils/scss/style.scss'
+import { globalStyles } from '../src/utils/styles'
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export const wrapRootElement = ({ element }) => <I18nextProvider i18n={i18next}>{element}</I18nextProvider>
+export const wrapRootElement = ({ element }) => (
+  <>
+    <Global styles={globalStyles} />
+    <I18nextProvider i18n={i18next}>{element}</I18nextProvider>
+  </>
+)
