+++ src/components/indexpage-components/index.tsx
@@ -1,0 +1,17 @@
+import AboutSection from './AboutSection'
+import BannerSection from './BannerSection'
+import CompetencesSection from './CompetencesSection'
+import EducationSection from './EducationSection'
+import NewsSection from './NewsSection'
+import PartnersSection from './PartnersSection'
+import ProjectsSection from './ProjectsSection'
+
+export {
+  BannerSection as Banner,
+  CompetencesSection as Competences,
+  EducationSection as Education,
+  AboutSection as About,
+  NewsSection as News,
+  PartnersSection as Partners,
+  ProjectsSection as Projects
+}
