+++ src/utils/scss/_overrides.scss
@@ -6,18 +6,18 @@
 //
 // Color system
 //
-$primary: #960018 !default;
-$secondary: #232b2b !default;
-$primary-medium: #b24557 !default;
-$secondary-medium: #878b8b !default;
-$primary-light: #cf8b96 !default;
-$secondary-light: #c3c5c5 !default;
-$success: $green !default;
-$info: $cyan !default;
-$warning: $yellow !default;
-$danger: $red !default;
-$light: $gray-100 !default;
-$dark: $gray-800 !default;
+$primary: #960018;
+$secondary: #232b2b;
+$primary-medium: #b24557;
+$secondary-medium: #878b8b;
+$primary-light: #cf8b96;
+$secondary-light: #c3c5c5;
+$success: $green;
+$info: $cyan;
+$warning: $yellow;
+$danger: $red;
+$light: $gray-100;
+$dark: $gray-800;
 
 $gray-1: #25262a;
 $gray-2: #5c626e;
@@ -29,4 +29,4 @@ $gray-4: #edf0f5;
 //
 
 $font-family-sans-serif: 'Muli', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif,
-  'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji' !default;
+  'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
