+++ src/utils/scss/_site-blocks.scss
@@ -33,8 +33,6 @@
   left: 0;
   min-height: 100%;
   min-width: 100%;
-  //background-color: black;
-  //opacity: 0.5;
   z-index: 1;
 }
 
@@ -60,14 +58,11 @@
 }
 
 .intro-section {
-  //background: $white;
-  //background-size: cover;
   color: $white;
   position: relative;
   &,
   .container .row {
     height: calc(100vh - 2 * 8px - 29px - 96px);
-    //height: 100%;
     min-height: 570px;
   }
   &:before {
@@ -75,7 +70,7 @@
     position: absolute;
     height: 100%;
     width: 100%;
-    background: #222; //linear-gradient(90deg, #999 0%, #222 5%, #222 95%, #999 100%);
+    background: #222;
     opacity: 0.5;
     border-bottom-right-radius: 0px;
     background-size: cover;
@@ -126,14 +121,11 @@
   }
 }
 .intro-section-alt {
-  //background: $secondary;
-  //background-size: cover;
   color: $secondary;
   position: relative;
   &,
   .container .row {
     height: calc(100vh - 2 * 8px - 29px - 96px);
-    //height: 100%;
     min-height: 570px;
   }
   &:before {
@@ -141,7 +133,7 @@
     position: absolute;
     height: 100%;
     width: 100%;
-    background: #fff; //linear-gradient(90deg, #999 0%, #fff 5%, #fff 95%, #999 100%);
+    background: #fff;
     opacity: 0.5;
     border-bottom-right-radius: 0px;
     background-size: cover;
@@ -420,23 +412,6 @@
     .owl-next {
       right: 0;
     }
-    /*.owl-prev,
-		.owl-next {
-			position: absolute;
-			text-align: center;
-			bottom: 5px;
-			width: 40px;
-			height: 40px;
-			transform: translateY(-50%);
-			background: rgba(255,255,255,.5);
-			> span {
-				font-size: 1.5rem;
-				position: absolute;
-				top: 50%;
-				left: 50%;
-				transform: translate(-50%, -50%);
-			}
-		}*/
   }
   .owl-dots {
     position: absolute;
@@ -536,7 +511,6 @@
     margin: 0 auto;
     z-index: 1;
     margin-top: -33.94px;
-    // border: 1px solid red;
     background: #ccc;
     > span {
       width: 96px;
@@ -606,11 +580,11 @@
     -webkit-box-direction: normal !important;
     -ms-flex-direction: column !important;
     flex-direction: column !important;
-    .course-1-main {
-    }
+
     figure {
       position: relative;
     }
+
     .category {
       background: $primary;
 
@@ -621,6 +595,7 @@
         margin-bottom: 0;
       }
     }
+
     .price {
       position: absolute;
       bottom: 0;
@@ -630,6 +605,7 @@
       background: $secondary;
       color: $white;
     }
+
     .course-1-content {
       padding: 20px 40px 0px 40px;
       text-align: center;
@@ -638,6 +614,7 @@
         font-size: 18px;
         color: $black;
       }
+
       .rating {
         text-align: center;
         > span {
@@ -645,11 +622,13 @@
           margin: 2px;
         }
       }
+
       .desc {
         font-size: 15px;
         color: gray;
       }
     }
+
     .course-1-footer {
       text-align: center !important;
       margin-top: auto !important;
@@ -661,12 +640,15 @@
 .education-1-item {
   box-shadow: 0 2px 10px 0 rgba($black, 0.1);
   background-color: $white;
+
   figure {
     position: relative;
     margin-bottom: 5;
   }
+
   .category {
     background: $primary;
+
     &.bg-primary {
       background: $primary;
       border-bottom: 10px solid $secondary;
@@ -678,12 +660,14 @@
     }
 
     padding: 20px 20px 15px 20px;
+
     h3 {
       color: $white;
       font-size: 20px;
       margin-bottom: 0;
     }
   }
+
   .price {
     position: absolute;
     bottom: 0;
@@ -693,9 +677,11 @@
     background: $secondary;
     color: $white;
   }
+
   .education-1-content {
     padding: 20px 30px;
     text-align: center;
+
     .header-left {
       position: absolute;
       left: 0;
@@ -704,19 +690,23 @@
       background: $secondary;
       color: $white;
     }
+
     h2 {
       text-align: center;
       margin: 0 0 10px 0;
       font-size: 18px;
       color: $black;
     }
+
     .rating {
       text-align: center;
+
       > span {
         display: inline-block;
         margin: 2px;
       }
     }
+
     .desc {
       text-align: center;
       font-size: 15px;
@@ -732,9 +722,11 @@
 .ftco-counter {
   overflow-x: hidden;
   position: relative;
+
   @include media-breakpoint-down(lg) {
     background-position: top center !important;
   }
+
   .img-video {
     width: 100%;
     display: block;
@@ -742,6 +734,7 @@
       height: 300px;
     }
   }
+
   .icon {
     span {
       font-size: 50px;
@@ -749,12 +742,14 @@
       line-height: 1;
     }
   }
+
   .block-18 {
     display: block;
     width: 100%;
     text-align: center;
     margin-bottom: 30px;
   }
+
   .text {
     strong.number {
       font-weight: 600;
@@ -762,6 +757,7 @@
       color: $primary;
       display: block;
     }
+
     span {
       display: block;
       font-size: 16px;
@@ -778,6 +774,7 @@
       }
     }
   }
+
   .counter-wrap {
     @include media-breakpoint-down(sm) {
       margin-bottom: 20px;
@@ -803,12 +800,12 @@ iframe.fb_iframe_widget_lift,
   .white-logo {
     -webkit-filter: brightness(0) invert(1);
     filter: brightness(0) invert(1);
-    //height: 70px;
     width: 200px;
     padding: 20px;
     margin: 10px 10px;
     border: 1px dotted rgba($white, 0.5);
   }
+
   column-count: 1;
   @include media-breakpoint-only(sm) {
     column-count: 2;
@@ -840,8 +837,6 @@ iframe.fb_iframe_widget_lift,
     object-fit: contain;
     height: 100%;
     width: 100%;
-    //height: 70px;
-    //width: 200px;
     padding: 20px;
     margin: 10px 10px;
     border: 1px dotted rgba($white, 0.5);
@@ -874,6 +869,7 @@ iframe.fb_iframe_widget_lift,
   background: $secondary;
   font-size: 14px;
   color: $white;
+
   .footer-heading {
     font-size: 16px;
     color: $white;
@@ -882,11 +878,13 @@ iframe.fb_iframe_widget_lift,
       border-bottom: 3px solid $primary;
     }
   }
+
   .copyright {
     color: $white;
     padding-top: 7rem;
     text-align: center;
   }
+
   .footer-column {
     padding-left: 3rem;
     padding-right: 3rem;
@@ -898,6 +896,7 @@ iframe.fb_iframe_widget_lift,
   background-position: center center;
   background-repeat: no-repeat;
   position: relative;
+
   &:before {
     position: absolute;
     content: '';
@@ -909,6 +908,7 @@ iframe.fb_iframe_widget_lift,
     right: 0;
     bottom: 0;
   }
+
   > .container {
     position: relative;
     z-index: 2;
@@ -917,6 +917,7 @@ iframe.fb_iframe_widget_lift,
       color: $white;
     }
   }
+
   .form-control {
     height: auto !important;
   }
@@ -930,6 +931,7 @@ iframe.fb_iframe_widget_lift,
   & a:not(.btn) {
     color: $primary-light;
     font-weight: 500;
+
     &:hover {
       color: $white;
       text-decoration: underline;
@@ -941,6 +943,7 @@ iframe.fb_iframe_widget_lift,
   & a {
     color: $gray-500;
     font-weight: 500;
+
     &:hover {
       color: $white;
       text-decoration: underline;
@@ -952,6 +955,7 @@ iframe.fb_iframe_widget_lift,
   & a {
     color: $primary-medium;
     font-weight: 500;
+
     &:hover {
       color: $primary-light;
       text-decoration: none;
@@ -963,6 +967,7 @@ iframe.fb_iframe_widget_lift,
   & a {
     color: $black;
     font-weight: 500;
+
     &:hover {
       color: $primary;
       text-decoration: none;
@@ -988,6 +993,7 @@ iframe.fb_iframe_widget_lift,
     color: $white;
     transform: translate(-50%, -50%);
     border-radius: 50%;
+
     > span {
       position: absolute;
       top: 50%;
@@ -995,6 +1001,7 @@ iframe.fb_iframe_widget_lift,
       transform: translate(-50%, -50%);
     }
   }
+
   &:hover {
     box-shadow: 0 5px 30px -10px rgba($black, 0.3);
   }
